backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Vector DB
chromadb==0.4.18

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
google-generativeai==0.3.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.3
python-magic==0.4.27
tiktoken==0.5.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2

# Monitoring
python-json-logger==2.0.7
